Clean up a bit
